src/processors/archive_processor/requirements.txt
patool==3.1.3
py7zr==0.22.0